Separate independent types into a distinct module
