+++ ponzu-back/Cargo.toml
@@ -7,8 +7,6 @@ edition = "2021"
 actix-web = "4"
 dotenv = "0.15.0"
 serde = "1.0.217"
-serde_json = "1.0.137"
 mongodb = "3.2.0"
 colored = "3.0.0"
 futures = "0.3.31"
-
