+++ ponzu-back/src/models/mod.rs
@@ -1,13 +1,11 @@
 // Exposed modules
 pub mod anime;
-pub mod app_error;
-pub mod app_state;
+pub mod character;
 pub mod club;
-pub mod error_response;
 pub mod genre;
 pub mod links;
 pub mod manga;
+pub mod person;
 pub mod producer;
 pub mod review;
-pub mod title_meta;
 pub mod user;
